refactor: flatten three single-owner architecture layers
Collapse the one-subclass Google handler base into its remaining consumer.
Remove the CLI team-plan identity aliases in favor of their shared owner.
Inline the proposal badge helper into its only production caller.
