+++ src/agent/modelHandlers/google/modelHandlerGoogleInteractions.ts
@@ -10,10 +10,12 @@ import {
   type Part,
   type Stream,
 } from '@google/genai';
+import { ReasoningEffort } from 'llm-zoo';
 
 // Local imports
 import { logProgressStatus } from '@agent/trace';
 import type { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState';
+import { ModelHandler } from '@agent/modelHandlers/ModelHandler';
 import { reportMediaAttachmentFailure } from '@agent/modelHandlers/support/mediaAttachmentPolicy';
 import { parseToolInputAsObject } from '@agent/core/flows/toolUseRound/toolCallParsing';
 import type { NormalizedUsage } from '@agent/types/NormalizedUsage';
@@ -25,8 +27,10 @@ import type {
   CreateResponseResult,
   ExtractResponseResult,
   GoogleToolCall,
+  ModelCredentialSelection,
   TokenCountOptions,
 } from '@agent/types/ModelHandlerContracts';
+import type { MediaEntry } from '@agent/utils/mediaTypes';
 import {
   attachManualRetryOnlyError,
   attachPartialText,
@@ -43,11 +47,18 @@ import type {
   ToolResult,
 } from '@shared/schemas/toolResult';
 import { filterNotNull, generateShortId, isNonEmptyString } from '@utils/core';
-import { joinNonEmpty } from '@utils/text/stringUtils';
+import { getShortDisplayPath } from '@utils/files';
+import { joinNonEmpty, pluralize } from '@utils/text/stringUtils';
 import { getConfig } from '@utils/config/configUtils';
 
 // Local file imports
-import { GoogleModelHandlerBase } from './GoogleModelHandlerBase';
+import {
+  resolveGoogleClient,
+  uploadGoogleMediaEntries,
+  type GoogleClientCache,
+  type GoogleMediaSource,
+} from './googleHandlerShared';
+import { tagGoogleSdkError } from './googleSdkError';
 import {
   computeGoogleInteractionsPrice,
   normalizeGoogleInteractionsUsage,
@@ -64,7 +75,6 @@ import {
   loadAttachmentBuffer,
 } from '../utils/toolAttachmentUtils';
 import { convertGoogleToolSchema, toGoogleTools } from '../toolConversion';
-import type { GoogleMediaSource } from './googleHandlerShared';
 
 // Interactions SDK aliases (public surface; the SDK re-exports these under the
 // `Interactions` namespace — the internal `_2`-suffixed types are not exported).
@@ -329,13 +339,13 @@ interface PendingBackgroundInteraction {
   readonly deadlineAtMs: number;
 }
 
-export class ModelHandlerGoogleInteractions extends GoogleModelHandlerBase<
+export class ModelHandlerGoogleInteractions extends ModelHandler<
   Step,
   Usage | null,
   GoogleToolCall,
+  GoogleGenAI,
   GoogleGenAIInteraction,
-  Content,
-  ThinkingLevel
+  Content
 > {
   // ===========================================================================
   // BACKGROUND mode (background:true + store:true, poll interactions.get)
@@ -509,6 +519,213 @@ export class ModelHandlerGoogleInteractions extends GoogleModelHandlerBase<
       );
       this.invalidateChain();
     }
+  }
+
+  // ===========================================================================
+  // Client cache + media pipeline
+  // ===========================================================================
+
+  /**
+   * Inline-vs-uploaded media threshold for the Google File API: 20 MiB.
+   */
+  private static readonly INLINE_MEDIA_LIMIT_BYTES = 20 * 1024 * 1024;
+
+  private googleClient: GoogleClientCache | null = null;
+
+  /** Whether the model can accept file attachments (image/video or native audio). */
+  protected supportsFileUploads(): boolean {
+    return (
+      this.capabilities.supportsVision || this.capabilities.supportsNativeAudio
+    );
+  }
+
+  /** Whether the model is a Gemini 3 variant (different thinking/media rules). */
+  protected isGemini3Model(): boolean {
+    return /^gemini-3[\.\-]/.test(this.config.fullName);
+  }
+
+  protected getInlineUploadLimitBytes(): number {
+    return ModelHandlerGoogleInteractions.INLINE_MEDIA_LIMIT_BYTES;
+  }
+
+  async getClient(
+    selection: ModelCredentialSelection = 'configured',
+  ): Promise<GoogleGenAI> {
+    const credential = await this.resolveClientCredential(selection);
+    return resolveGoogleClient({
+      sdkLabel: this.sdkLabel,
+      credential,
+      logger: this.logger,
+      cached: this.googleClient,
+      setCached: (cache) => {
+        this.googleClient = cache;
+      },
+      rememberRoute: (client, route, credentialSecret) =>
+        this.rememberClientCredentialRoute(client, route, credentialSecret),
+    });
+  }
+
+  override async refreshClient(
+    selection: ModelCredentialSelection = 'configured',
+  ): Promise<GoogleGenAI> {
+    this.googleClient = null;
+    return this.getClient(selection);
+  }
+
+  protected override get sdkErrorTagger() {
+    return tagGoogleSdkError;
+  }
+
+  override get supportsForcedToolChoice(): boolean {
+    return true;
+  }
+
+  /**
+   * Google passes the system prompt per call as `system_instruction` rather
+   * than storing it in `messages`, so the round flow must resupply it on every
+   * invocation.
+   */
+  override get requiresPerCallSystemPrompt(): boolean {
+    return true;
+  }
+
+  /**
+   * Gemini carries thought signatures across parallel function calls, which must
+   * be preserved by batching the results into a single follow-up: the handler
+   * rebuilds the model-generated turn (thought signatures + every function call)
+   * ahead of the results, in the order the model emitted them. The tool-use flow
+   * only records the assistant turn through the follow-up methods (see
+   * `ToolUseDispatchNode`), so splitting parallel calls across separate turns
+   * would lose the signature and reference a call absent from the transcript.
+   * Unconditional (not gated on `capabilities.supportsReasoning`) — see the base
+   * getter's doc comment (#7101 triage).
+   */
+  override get requiresBatchedParallelToolResults(): boolean {
+    return true;
+  }
+
+  /**
+   * Map the model's reasoning effort to a Gemini `thinking_level`.
+   *
+   * Returns `levels.low` (not `undefined`) for `NONE` so the API does not fall
+   * back to its default medium/high — that would defeat the user's intent.
+   * Gemini tops out at HIGH thinking, so xhigh/max both map to it; Gemini 3 Pro
+   * only supports low/high, so MEDIUM falls back to HIGH for Pro.
+   */
+  protected getThinkingLevel(): ThinkingLevel | undefined {
+    const { levels, labels } = this.thinkingLevelConfig;
+    const isGemini3 = this.isGemini3Model();
+
+    switch (this.capabilities.reasoningEffort) {
+      case ReasoningEffort.NONE:
+        if (isGemini3) {
+          this.logger.warn(
+            `Gemini 3 models can't fully disable thinking. Using thinking_level '${labels.low}'.`,
+          );
+        }
+        return levels.low;
+
+      case ReasoningEffort.LOW:
+        return levels.low;
+
+      case ReasoningEffort.MEDIUM:
+        if (isGemini3 && this.config.fullName.includes('-pro')) {
+          this.logger.debug(
+            `Gemini 3 Pro does not support ${labels.medium} thinking level. Using ${labels.high}.`,
+          );
+          return levels.high;
+        }
+        return levels.medium;
+
+      case ReasoningEffort.HIGH:
+      case ReasoningEffort.XHIGH:
+      case ReasoningEffort.MAX:
+        return levels.high;
+
+      default:
+        return undefined;
+    }
+  }
+
+  protected override async createMediaMessage(
+    mediaFiles: FileLocation[],
+  ): Promise<Content[]> {
+    if (!mediaFiles?.length || !this.supportsFileUploads()) {
+      return [];
+    }
+
+    const { entries, results } =
+      await this.mediaProcessor.loadEntries(mediaFiles);
+    this.mediaProcessor.logResults(results);
+
+    if (entries.length === 0) {
+      return [];
+    }
+
+    return this.uploadMediaEntries(entries);
+  }
+
+  /** Media blocks for the entries (inline when small enough, uploaded otherwise). */
+  protected async uploadMediaEntries(
+    entries: MediaEntry[],
+  ): Promise<Content[]> {
+    const insertedEntries: MediaEntry[] = [];
+    const media = await uploadGoogleMediaEntries<Content>(entries, {
+      getClient: () => this.getClient(),
+      inlineLimit: this.getInlineUploadLimitBytes(),
+      logger: this.logger,
+      buildMedia: (source, mimeType) => this.buildMedia(source, mimeType),
+      onInsertedEntry: (entry) => insertedEntries.push(entry),
+    });
+    this.setCreatedMediaEntriesForAttachmentLog(insertedEntries);
+    return media;
+  }
+
+  /** Labelled media for a round, or empty when there is nothing to attach. */
+  protected async buildLabelledMedia(
+    mediaFiles: FileLocation[] | undefined,
+    context: 'initial' | 'followUp',
+  ): Promise<Content[]> {
+    if (!mediaFiles?.length || !this.supportsFileUploads()) {
+      return [];
+    }
+
+    const media = await this.createMediaForRound(mediaFiles, context);
+    if (media.length === 0) {
+      return [];
+    }
+
+    const label = mediaFiles.map((loc) => getShortDisplayPath(loc)).join(', ');
+    const verb = context === 'initial' ? 'Attached' : 'Processing';
+    return [
+      this.textMedia(
+        `\n${verb} ${pluralize(mediaFiles.length, 'file')}: ${label}`,
+      ),
+      ...media,
+    ];
+  }
+
+  /**
+   * Follow-up for a SINGLE tool call, built by delegating to the batched
+   * path with a one-entry array so both call sites share one assembly.
+   */
+  async createToolUseFollowUpMessages(
+    _client: GoogleGenAI | undefined,
+    call: GoogleToolCall,
+    result: ToolResult,
+    attachments: ToolFileAttachment[],
+    workspaceState?: AgentWorkspaceState,
+    text?: string,
+  ): Promise<Step[]> {
+    if (!call.callId) {
+      throw new Error('Function call id is required for follow-up messages');
+    }
+
+    return this.createBatchedToolUseFollowUpMessages(
+      [{ call, result, attachments }],
+      workspaceState,
+      text,
+    );
   }
 
   // ===========================================================================
